parse response data and move extractQuestion to a hook.
