+++ services/app-web/src/hooks/index.ts
@@ -9,3 +9,4 @@ export { useScrollToPageTop } from './useScrollToPageTop'
 export { useTealium } from './useTealium'
 export { useTitle } from './useTitle'
 export { useStatePrograms } from './useStatePrograms'
+export { useQuestions } from './useQuestions'
